fix(ssl/tls) update ssl/tls client configuration
* switch to green mail for integration tests
