+++ pom.xml
@@ -5,7 +5,7 @@
 
 	<groupId>org.bonitasoft.connectors</groupId>
 	<artifactId>bonita-connector-email</artifactId>
-	<version>1.2.4-SNAPSHOT</version>
+	<version>1.3.0-SNAPSHOT</version>
 
 	<name>Bonita Email Connector</name>
 	<description>Email Connector for Bonita</description>
@@ -53,12 +53,12 @@
 		<!-- Test -->
 		<junit-jupiter-engine.version>5.6.2</junit-jupiter-engine.version>
 		<junit-platform-runner.version>1.6.2</junit-platform-runner.version>
-		<subethamail.version>3.1.5</subethamail.version>
+		<greenmail.version>1.6.0</greenmail.version>
 		<jacoco-maven-plugin.version>0.8.5</jacoco-maven-plugin.version>
 		<sonar-maven-plugin.version>3.7.0.1746</sonar-maven-plugin.version>
 		<mockito-core.version>3.5.13</mockito-core.version>
 		<assertj-core.version>3.16.1</assertj-core.version>
-		<awaitility.version>2.0.0</awaitility.version>
+		<logback-classic.version>1.2.3</logback-classic.version>
 
 		<!-- Maven plugins -->
 		<maven-compiler-plugin.version>3.8.1</maven-compiler-plugin.version>
@@ -143,9 +143,15 @@
 			<scope>test</scope>
 		</dependency>
 		<dependency>
-			<groupId>org.subethamail</groupId>
-			<artifactId>subethasmtp</artifactId>
-			<version>${subethamail.version}</version>
+			<groupId>com.icegreen</groupId>
+			<artifactId>greenmail</artifactId>
+			<version>${greenmail.version}</version>
+			<scope>test</scope>
+		</dependency>
+		<dependency>
+			<groupId>com.icegreen</groupId>
+			<artifactId>greenmail-junit5</artifactId>
+			<version>${greenmail.version}</version>
 			<scope>test</scope>
 		</dependency>
 		<dependency>
@@ -155,9 +161,9 @@
 			<scope>test</scope>
 		</dependency>
 		<dependency>
-			<groupId>org.awaitility</groupId>
-			<artifactId>awaitility</artifactId>
-			<version>${awaitility.version}</version>
+			<groupId>ch.qos.logback</groupId>
+			<artifactId>logback-classic</artifactId>
+			<version>${logback-classic.version}</version>
 			<scope>test</scope>
 		</dependency>
 	</dependencies>
@@ -230,6 +236,10 @@
 						<enableAssertions>true</enableAssertions>
 						<forkCount>1</forkCount>
 						<reuseForks>true</reuseForks>
+						<includes>
+						<include>**/*IT.java</include>
+						<include>**/*Test.java</include>
+						</includes>
 					</configuration>
 					<dependencies>
 						<dependency>
