+++ src/main/resources/email.properties
@@ -8,8 +8,8 @@ port.label=SMTP Port
 security.label=Security
 ssl.label=SSL
 starttls.label=STARTTLS
-autoTrustCertificate.label=Auto-trust SSL certificate
-autoTrustCertificate.description=When using a SSL connection, the SMTP server certificate is always trusted if true. 
+trustCertificate.label=Trust certificate
+trustCertificate.description=When using a SSL/TLS connection, the SMTP server certificate is always trusted if true. 
 authentication.label=Authentication
 username.label=Username
 password.label=Password
